Add factory pageView constructor to AutoTabsRouter
This is a breaking change as I honestly don't know how to keep the default constructor and still make it work. The two constructors don't share much in common besides the `routes` field and I didn't want to make a delegate although maybe I could. Factories are a bit confusing to me but feel free to offer any advice